% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{automate_data_cleaning}
\alias{automate_data_cleaning}
\title{Automation Data Cleaning for all columns with NA}
\usage{
automate_data_cleaning(df)
}
\arguments{
\item{df}{The data frame that need Data Cleaning (include outlier & NA value treated)}
}
\value{
Data frame that successfully cleaned
}
\description{
Automation Data Cleaning for all columns with NA
}
