% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{automation_knn_imputation}
\alias{automation_knn_imputation}
\title{Automation kNN imputation for all columns with NA}
\usage{
automation_knn_imputation(df, ignore_col = NULL)
}
\arguments{
\item{df}{Data frame}

\item{ignore_col}{Character vector of columns to ignore}
}
\value{
Data frame with selected columns imputed
}
\description{
Automation kNN imputation for all columns with NA
}
