% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_p_categorical.R
\name{backward_p_lr}
\alias{backward_p_lr}
\title{Make ensemble prediction by averaging probabilities}
\usage{
backward_p_lr(
  df,
  train_index,
  target_col,
  positive,
  positive_weight,
  negative_weight
)
}
\arguments{
\item{df}{The data frame those index fall into}

\item{train_index}{The index for ensemble/not ensemble train}

\item{target_col}{The target column that need to be predicted}

\item{positive}{Positive value of target_col}

\item{positive_weight}{If class imbalance within data set, add weight based on portion}

\item{negative_weight}{If class imbalance within data set, add weight based on portion}
}
\value{
A list of trained ensemble Logistic Model Based on List of Index you provide
}
\description{
Make ensemble prediction by averaging probabilities
}
