% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_p_numerical.R
\name{backward_p_mlr}
\alias{backward_p_mlr}
\title{Backward p-value for Multi-Linear Regression Model}
\usage{
backward_p_mlr(df, train_index, target_col)
}
\arguments{
\item{df}{The data frame those index fall into}

\item{train_index}{The index for ensemble/not ensemble train}

\item{target_col}{The target column that need to be predicted}
}
\value{
A list of trained emseble Logistic Model Based on List of Index you provide
}
\description{
Backward p-value for Multi-Linear Regression Model
}
