% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{check_class_imbalance}
\alias{check_class_imbalance}
\title{Check class imbalance based on target}
\usage{
check_class_imbalance(df, index_list, target_col, positive, negative)
}
\arguments{
\item{df}{The data frame those index fall into}

\item{index_list}{The list of index you want to check if fall into class imbalance}

\item{target_col}{The column that used to check class imbalance}

\item{positive}{Positive value of target_col}

\item{negative}{Negative value of target_col}
}
\value{
No Return, but will output result
}
\description{
Check class imbalance based on target
}
