% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{check_model_performance}
\alias{check_model_performance}
\title{Check model performance: accuracy, TPR, TNR, F1}
\usage{
check_model_performance(
  predict_prob,
  threshold,
  positive,
  negative,
  df,
  test_index,
  target_col
)
}
\arguments{
\item{predict_prob}{Numeric vector or list of predictions}

\item{threshold}{Threshold to classify positive}

\item{positive}{Positive label}

\item{negative}{Negative label}

\item{df}{Data frame}

\item{test_index}{Indices used for evaluation}

\item{target_col}{Target column}
}
\value{
List with accuracy, tpr, tnr, F1
}
\description{
Check model performance: accuracy, TPR, TNR, F1
}
