% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{check_multicollinearity}
\alias{check_multicollinearity}
\title{Check multicollinearity via correlation matrix}
\usage{
check_multicollinearity(variables_df, threshold = 0.8)
}
\arguments{
\item{variables_df}{Data frame of numeric variables}

\item{threshold}{Correlation threshold to flag strong correlation}
}
\value{
Printed messages of strongly correlated pairs
}
\description{
Check multicollinearity via correlation matrix
}
