% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{check_na_zero}
\alias{check_na_zero}
\title{Check NA and zero values in numeric columns}
\usage{
check_na_zero(df)
}
\arguments{
\item{df}{Data frame to check}
}
\value{
Zero Count
}
\description{
Check NA and zero values in numeric columns
}
