% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_numerical_with_mlr.R
\name{create_stack_model_mlr}
\alias{create_stack_model_mlr}
\title{Create Stack Model for MLR}
\usage{
create_stack_model_mlr(df, target_col, test_index, model_list)
}
\arguments{
\item{df}{Data frame that you use to make prediction}

\item{target_col}{That column that you want to make prediction for}

\item{test_index}{The test index selected earlier}

\item{model_list}{The list of ensemble model you made ealier}
}
\value{
stacked model you generate from ensemble models
}
\description{
Create Stack Model for MLR
}
