% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross validation wrapper for LR / RF / C5.0 / rpart}
\usage{
cross_validation(
  df,
  index_list,
  positive,
  negative,
  positive_weight,
  negative_weight,
  target_col,
  model_type,
  seed_num = 888,
  threshold = 0.5
)
}
\arguments{
\item{df}{Data frame}

\item{index_list}{List of folds (indices)}

\item{positive}{Positive label}

\item{negative}{Negative label}

\item{positive_weight}{Positive class weight}

\item{negative_weight}{Negative class weight}

\item{target_col}{Target column}

\item{model_type}{"lr", "rf", "c5", or "rpart"}

\item{seed_num}{Seed}

\item{threshold}{Classification threshold}
}
\value{
Printed fold-wise and overall performance
}
\description{
Cross validation wrapper for LR / RF / C5.0 / rpart
}
