% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_numerical.R
\name{emsemble_result_with_weight_numerical}
\alias{emsemble_result_with_weight_numerical}
\title{Calculate How Each Ensemble Model should be weighted in Overall Model with RMSE}
\usage{
emsemble_result_with_weight_numerical(
  model_list,
  df,
  index,
  weight_list,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{The logistic regression model you create with your ensemble list}

\item{df}{Data frame that you use to make prediction}

\item{index}{The index you wanna try with this model (usually test & val index)}

\item{weight_list}{The weight you get for each ensemble based on RMSE}

\item{target_treatment}{If target column has been transformed, what method it use}
}
\value{
How each model should contribute to overall prediction
}
\description{
Calculate How Each Ensemble Model should be weighted in Overall Model with RMSE
}
