% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{ensemble_result_with_weight_categorical}
\alias{ensemble_result_with_weight_categorical}
\title{Support Logistic Regression, random forest, C5.0, rpart}
\usage{
ensemble_result_with_weight_categorical(
  model_list,
  df_list,
  index,
  weight_list,
  positive,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{The list of model, support logistic regression, C5.0, rpart, random forest}

\item{df_list}{The list of data frame align with model_list}

\item{index}{The index you wanna try with this model (usually test & val index)}

\item{weight_list}{The weight you get for each ensemble}

\item{positive}{Target's positive value}

\item{target_treatment}{If target feature is transformed in feature transformation (default="none")}
}
\value{
List of prediction (in probability) made with ensemble Logistic Regression model
}
\description{
Support Logistic Regression, random forest, C5.0, rpart
}
