% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{ensemble_train_partition_no_target}
\alias{ensemble_train_partition_no_target}
\title{Create Ensmeble Data Set with No Target (for numerical prediction)}
\usage{
ensemble_train_partition_no_target(df, train_index, sub_num, sub_portion)
}
\arguments{
\item{df}{Data frame that used for training partition}

\item{train_index}{The overall training index that you gonna selected from}

\item{sub_num}{Number of ensemble list}

\item{sub_portion}{Portion of overall training you want to select as sub-list}
}
\value{
A list contain sub_num number of training index
}
\description{
Create Ensmeble Data Set with No Target (for numerical prediction)
}
