% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{ensemble_weight_F1}
\alias{ensemble_weight_F1}
\title{Get Each Ensemble Model Weight with F1}
\usage{
ensemble_weight_F1(
  model_list,
  df_list,
  test_index,
  best_threshold,
  target_col,
  positive,
  negative
)
}
\arguments{
\item{model_list}{The list of model, support logistic regression, C5.0, rpart, random forest}

\item{df_list}{The list of data frame align with model_list}

\item{test_index}{The index of 25\% testing}

\item{best_threshold}{The best threshold you get after run function "find_best_threshold"}

\item{target_col}{The target column that need to make prediction}

\item{positive}{Target positive value}

\item{negative}{Target negative value}
}
\value{
Each Ensemble Model's Weight
}
\description{
Get Each Ensemble Model Weight with F1
}
