% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{find_best_k}
\alias{find_best_k}
\title{Find best k for kNN imputation based on RMSE}
\usage{
find_best_k(max_k, test_k_index, df_distance, not_na_index, target_feature)
}
\arguments{
\item{max_k}{Number of k you wanna test}

\item{test_k_index}{The index you wanna use to test best k}

\item{df_distance}{The euclidean distance matrix build before}

\item{not_na_index}{The not_na_index you get from previous function}

\item{target_feature}{The feature need to find best k}
}
\value{
Smallest_k used for kNN imputation
}
\description{
Find best k for kNN imputation based on RMSE
}
