% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{find_best_threshold}
\alias{find_best_threshold}
\title{Find best threshold to maximize F1}
\usage{
find_best_threshold(
  predict_prob,
  df,
  test_index,
  target_col,
  positive,
  negative
)
}
\arguments{
\item{predict_prob}{Numeric vector of probabilities}

\item{df}{Data frame}

\item{test_index}{Indices used}

\item{target_col}{Target column name}

\item{positive}{Positive label}

\item{negative}{Negative label}
}
\value{
Best threshold (numeric)
}
\description{
Find best threshold to maximize F1
}
