% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_categorical_with_glm.R
\name{generate_stack_df}
\alias{generate_stack_df}
\title{Helper Function that Generate Data Frame Suitable for Stack Model Construction}
\usage{
generate_stack_df(model_list, df_list, index, positive)
}
\arguments{
\item{model_list}{List of Model Generated Before, support glm, RF, C5, rpart}

\item{df_list}{List of data frame that aligned with model list}

\item{index}{The index you wanna use to train this model, could be test or validation}

\item{positive}{Target's positive value}
}
\value{
A fitted stack data frame
}
\description{
Helper Function that Generate Data Frame Suitable for Stack Model Construction
}
