% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{initialize_distance_find_best_k}
\alias{initialize_distance_find_best_k}
\title{Initialize distance matrix for kNN imputation}
\usage{
initialize_distance_find_best_k(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Euclidean distance matrix for numeric columns without NA
}
\description{
Initialize distance matrix for kNN imputation
}
