% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{initialize_not_na_index}
\alias{initialize_not_na_index}
\title{Get indices of non-NA values in a feature}
\usage{
initialize_not_na_index(target_feature)
}
\arguments{
\item{target_feature}{Specific feature that need to find not na value}
}
\value{
Target_feature's all index that are not na
}
\description{
Get indices of non-NA values in a feature
}
