% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{initialize_test_k_index}
\alias{initialize_test_k_index}
\title{Sample indices to tune k for kNN imputation}
\usage{
initialize_test_k_index(test_proportion, not_na_index)
}
\arguments{
\item{test_proportion}{Proportion you want to select from not_na_index to test best k}

\item{not_na_index}{The not_na_index you get from previous function}
}
\value{
The index you wanna use to test best k
}
\description{
Sample indices to tune k for kNN imputation
}
