% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{IQR_outlier}
\alias{IQR_outlier}
\title{IQR-based outlier treatment with capping at 5\% / 95\% quantiles}
\usage{
IQR_outlier(df, variables)
}
\arguments{
\item{df}{Data frame}

\item{variables}{Character vector of numeric variable names}
}
\value{
Data frame with outliers capped
}
\description{
IQR-based outlier treatment with capping at 5\% / 95\% quantiles
}
