% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{k_stratified_cv}
\alias{k_stratified_cv}
\title{Create k-fold stratified CV index list from train indices}
\usage{
k_stratified_cv(k, df, target_col, train_index, positive, negative)
}
\arguments{
\item{k}{Number of folds}

\item{df}{Data frame}

\item{target_col}{Target column}

\item{train_index}{Train indices}

\item{positive}{Positive class}

\item{negative}{Negative class}
}
\value{
List of length k with indices
}
\description{
Create k-fold stratified CV index list from train indices
}
