% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{kNN_Imputation}
\alias{kNN_Imputation}
\title{Perform kNN imputation for a single feature}
\usage{
kNN_Imputation(df, smallest_k, target_feature, df_distance)
}
\arguments{
\item{df}{Data frame}

\item{smallest_k}{Chosen k}

\item{target_feature}{Numeric vector with NA}

\item{df_distance}{Distance matrix}
}
\value{
Imputed numeric vector
}
\description{
Perform kNN imputation for a single feature
}
