% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{make_ensemble_predict}
\alias{make_ensemble_predict}
\title{Make ensemble prediction by averaging probabilities}
\usage{
make_ensemble_predict(
  model_list,
  df_list,
  test_index,
  positive,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{List of models (glm, randomForest, C5.0, rpart)}

\item{df_list}{List of data frames (or single data frame)}

\item{test_index}{Indices to predict}

\item{positive}{Positive class label}

\item{target_treatment}{Transformation flag (unused in your current code)}
}
\value{
Numeric vector of ensemble probabilities
}
\description{
Make ensemble prediction by averaging probabilities
}
