% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{make_ensemble_predict_categorical}
\alias{make_ensemble_predict_categorical}
\title{This Function Support Logistic Regression, RandomForest, C5.0 & rpart}
\usage{
make_ensemble_predict_categorical(
  model_list,
  df_list,
  test_index,
  positive,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{List of logistic regression model}

\item{df_list}{The list of data frame that aligned with model_list}

\item{test_index}{The list of index for ensemble train}

\item{positive}{Positive value of target_col}

\item{target_treatment}{If target feature is transformed in feature transformation (default="none")}
}
\value{
No Return, but will output result
}
\description{
This Function Support Logistic Regression, RandomForest, C5.0 & rpart
}
