% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_numerical.R
\name{make_ensemble_predict_numerical}
\alias{make_ensemble_predict_numerical}
\title{Make Ensemble Prediction for Numcerical Prediction}
\usage{
make_ensemble_predict_numerical(
  model_list,
  df,
  test_index,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{List of logistic regression model}

\item{df}{The data frame those index fall into}

\item{test_index}{The list of index for ensemble train}

\item{target_treatment}{If transformation occured in feature transformation}
}
\value{
Mean Prediction made by number of ensemble's logistic model
}
\description{
Make Ensemble Prediction for Numcerical Prediction
}
