% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{median_imputation}
\alias{median_imputation}
\title{Median imputation for selected columns}
\usage{
median_imputation(df, target_col)
}
\arguments{
\item{df}{Data frame}

\item{target_col}{Character vector of column names to impute}
}
\value{
Data frame with NA imputed by median
}
\description{
Median imputation for selected columns
}
