% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{replace_na_with_zero}
\alias{replace_na_with_zero}
\title{Replace zeros with NA except in specified columns}
\usage{
replace_na_with_zero(df, ignore_cols)
}
\arguments{
\item{df}{Data frame}

\item{ignore_cols}{Character vector of column names whose zeros should NOT be replaced}
}
\value{
Data frame with zeros replaced by NA in selected numeric columns
}
\description{
Replace zeros with NA except in specified columns
}
