% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{replace_zero_with_na}
\alias{replace_zero_with_na}
\title{Replace zeros with NA except in specified columns}
\usage{
replace_zero_with_na(df, ignore_cols)
}
\arguments{
\item{df}{Data frame with zero value}

\item{ignore_cols}{Zero value that does not need to be changed into NA}
}
\value{
Data frame with designated columns' zero value replaced into na
}
\description{
Replace zeros with NA except in specified columns
}
