% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{reverse_num}
\alias{reverse_num}
\title{Reverse Numeric Transformation}
\usage{
reverse_num(x, treatment)
}
\arguments{
\item{x}{Numeric vector.}

\item{treatment}{One of \code{"log"}, \code{"log1p"}, \code{"sqrt"},
\code{"square"}, \code{"none"}.}
}
\value{
Numeric vector with inverse transformation applied.
}
\description{
Applies the inverse of a transformation to a numeric vector.
}
