% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_categorical_with_glm.R
\name{stack_model_lr}
\alias{stack_model_lr}
\title{Purpose is use a list of model to train a stack model}
\usage{
stack_model_lr(model_list, df_list, test_index, positive, target_col)
}
\arguments{
\item{model_list}{List of Model Generated Before, support glm, RF, C5, rpart}

\item{df_list}{List of data frame that aligned with model list}

\item{test_index}{The index you use prepare to tune model}

\item{positive}{Target's positive value}

\item{target_col}{The feature you wanna make prediction}
}
\value{
A stack model made with a list of model
}
\description{
Support logistic regression, random forest, C5.0 & rpart
}
