% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_categorical_with_glm.R
\name{stack_model_predict_lr}
\alias{stack_model_predict_lr}
\title{This function's purpose is to use the stack model you build earlier to make prediction}
\usage{
stack_model_predict_lr(
  stack_model,
  model_list,
  df_list,
  index,
  positive,
  negative,
  target_col,
  threshold = 0.5
)
}
\arguments{
\item{stack_model}{The stack_model you build earlier}

\item{model_list}{A list of model, support logistic regression, random forest, C5.0 & rpart}

\item{df_list}{A list of data frame that aligned with model list}

\item{index}{The index you wanna test this model, could be test & validation}

\item{positive}{The positive value of your target}

\item{negative}{The negative value of your target}

\item{target_col}{The feature you wanna make prediction}

\item{threshold}{I set to reasonable 0.5, but you can change this after threshold tune}
}
\value{
Output the overall performance of stack model
}
\description{
This function's purpose is to use the stack model you build earlier to make prediction
}
