% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{three_set_partition}
\alias{three_set_partition}
\title{Split data into train, test, and validation with stratified sampling}
\usage{
three_set_partition(
  df,
  seed_num,
  target_col,
  train_portion,
  test_portion,
  positive,
  negative
)
}
\arguments{
\item{df}{Data frame}

\item{seed_num}{Seed}

\item{target_col}{Target column name}

\item{train_portion}{Train proportion}

\item{test_portion}{Test proportion (of full data)}

\item{positive}{Positive class label}

\item{negative}{Negative class label}
}
\value{
List with train_index, test_index, validation_index
}
\description{
Split data into train, test, and validation with stratified sampling
}
