% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{z_score_outlier}
\alias{z_score_outlier}
\title{Z-score based outlier treatment (±3 SD)}
\usage{
z_score_outlier(df, variable)
}
\arguments{
\item{df}{The data frame that has feature need outlier treatment}

\item{variable}{Variables that need outlier treatment & normally}
}
\value{
Data frame with extreme values capped at mean ± 3*sd
}
\description{
Z-score based outlier treatment (±3 SD)
}
