% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{automation_knn_imputation}
\alias{automation_knn_imputation}
\title{Automation kNN imputation for all columns with NA}
\usage{
automation_knn_imputation(df, test_por = 0.1, max_k = 20, ignore_col = NULL)
}
\arguments{
\item{df}{The data frame that has NA need kNN imputation treatment}

\item{test_por}{Proportion of the dataset used for k testing (default = 10\%)}

\item{max_k}{Maximum number of k test for best k (default=20)}

\item{ignore_col}{Columns name that does not need its NA be treated (default=NULL)}
}
\value{
Data frame with designated column successfully impute
}
\description{
Automation kNN imputation for all columns with NA
}
