% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper.R
\name{backward_p_lr}
\alias{backward_p_lr}
\title{Logistic regression with backward elimination based on p-values}
\usage{
backward_p_lr(
  df,
  train_index,
  target_col,
  positive,
  positive_weight,
  negative_weight
)
}
\arguments{
\item{df}{Data frame}

\item{train_index}{Numeric vector or list of vectors}

\item{target_col}{Target column name}

\item{positive}{Positive class label}

\item{positive_weight}{Weight for positive class}

\item{negative_weight}{Weight for negative class}
}
\value{
glm object or list of glm objects
}
\description{
Logistic regression with backward elimination based on p-values
}
