% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{check_class_imbalance}
\alias{check_class_imbalance}
\title{Check class balance for indices or list of index sets}
\usage{
check_class_imbalance(df, index_list, target_col, positive, negative)
}
\arguments{
\item{df}{Data frame}

\item{index_list}{Numeric vector or list of numeric vectors}

\item{target_col}{Target column}

\item{positive}{Positive class label}

\item{negative}{Negative class label}
}
\value{
Printed messages
}
\description{
Check class balance for indices or list of index sets
}
