% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{check_model_performance}
\alias{check_model_performance}
\title{Check model performance: accuracy, TPR, TNR, F1}
\usage{
check_model_performance(
  predict_prob,
  threshold,
  positive,
  negative,
  df,
  test_index,
  target_col
)
}
\arguments{
\item{predict_prob}{The prediction result (in probability/response)}

\item{threshold}{Above what number is positive}

\item{positive}{Target's positive value}

\item{negative}{Target's negative value}

\item{df}{The data frame you use to make prediction}

\item{test_index}{The index you select for test}

\item{target_col}{Your target name}
}
\value{
A list of result include accuracy, tpr, tnr, F1 value and output result
}
\description{
Check model performance: accuracy, TPR, TNR, F1
}
