% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{check_multicollinearity}
\alias{check_multicollinearity}
\title{Check multicollinearity via correlation matrix}
\usage{
check_multicollinearity(variables_df, threshold = 0.8)
}
\arguments{
\item{variables_df}{Variables selected for multicollinearity checking (has to be numerical)}

\item{threshold}{Above what correlation value is considered as strong correlation}
}
\value{
Nothing will return but will tell you what features are highly correlated
}
\description{
Check multicollinearity via correlation matrix
}
