% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_categorical.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross Validation with k Stratified Fold Index}
\usage{
cross_validation(
  df,
  index_list,
  positive,
  negative,
  positive_weight,
  negative_weight,
  target_col,
  model_type,
  threshold = 0.5
)
}
\arguments{
\item{df}{Data Frame you wanna test with}

\item{index_list}{The index list generated earlier (with function k_stratified_cv)}

\item{positive}{target_col's positive value}

\item{negative}{target_col's negative value}

\item{positive_weight}{The portion of positive of overall value}

\item{negative_weight}{The portion of negative of overall value}

\item{target_col}{The column need to be predicted}

\item{model_type}{This function support "logistic regression", "random forest","c5.0","decision tree"}

\item{threshold}{Above threshold is considered positive, vice versa (default=0.5)}
}
\value{
k stratified index list
}
\description{
Cross Validation with k Stratified Fold Index
}
