% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{emsemble_result_with_weight}
\alias{emsemble_result_with_weight}
\title{Weighted ensemble prediction}
\usage{
emsemble_result_with_weight(
  model_list,
  df_list,
  index,
  weight_list,
  positive,
  target_treatment = "none"
)
}
\arguments{
\item{model_list}{List of models}

\item{df_list}{List of data frames or single data frame}

\item{index}{Indices to predict}

\item{weight_list}{List of weights}

\item{positive}{Positive label}

\item{target_treatment}{Transformation flag}
}
\value{
Numeric vector of weighted ensemble probabilities
}
\description{
Weighted ensemble prediction
}
