% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{ensemble_train_partition}
\alias{ensemble_train_partition}
\title{Create ensemble training subsets with stratified sampling & bootstrapping}
\usage{
ensemble_train_partition(
  df,
  seed_num,
  train_index,
  sub_num,
  sub_portion,
  target_col,
  positive,
  negative
)
}
\arguments{
\item{df}{Data frame}

\item{seed_num}{Seed}

\item{train_index}{Overall train indices}

\item{sub_num}{Number of subsets}

\item{sub_portion}{Proportion of each class to sample}

\item{target_col}{Target column}

\item{positive}{Positive class label}

\item{negative}{Negative class label}
}
\value{
List of index vectors
}
\description{
Create ensemble training subsets with stratified sampling & bootstrapping
}
