% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{find_best_k}
\alias{find_best_k}
\title{Find best k for kNN imputation based on RMSE}
\usage{
find_best_k(max_k, test_k_index, df_distance, not_na_index, target_feature)
}
\arguments{
\item{max_k}{Maximum k to search}

\item{test_k_index}{Indices used for tuning}

\item{df_distance}{Distance matrix}

\item{not_na_index}{Non-NA indices}

\item{target_feature}{Numeric feature}
}
\value{
Best k (integer)
}
\description{
Find best k for kNN imputation based on RMSE
}
