% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_categorical.R
\name{find_best_threshold}
\alias{find_best_threshold}
\title{Reminder: target must be numerical value}
\usage{
find_best_threshold(
  predict_prob,
  df,
  test_index,
  target_col,
  positive,
  negative
)
}
\arguments{
\item{predict_prob}{The prediction result (in probability/response)}

\item{df}{The data frame you use to make prediction}

\item{test_index}{The index you select for test}

\item{target_col}{Your target name}

\item{positive}{Target positive value}

\item{negative}{Target negative value}
}
\value{
Best Threshold with optimal F1 value
}
\description{
Reminder: target must be numerical value
}
