% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{initialize_distance_find_best_k}
\alias{initialize_distance_find_best_k}
\title{Initialize distance matrix for kNN imputation}
\usage{
initialize_distance_find_best_k(df)
}
\arguments{
\item{df}{Data frame that need to be treated}
}
\value{
The euclidean distance list that will be used
}
\description{
Initialize distance matrix for kNN imputation
}
