% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{initialize_not_na_index}
\alias{initialize_not_na_index}
\title{Get indices of non-NA values in a feature}
\usage{
initialize_not_na_index(target_feature)
}
\arguments{
\item{target_feature}{Numeric vector}
}
\value{
Integer vector of indices without NA
}
\description{
Get indices of non-NA values in a feature
}
