% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_imputation.R
\name{initialize_test_k_index}
\alias{initialize_test_k_index}
\title{Sample indices to tune k for kNN imputation}
\usage{
initialize_test_k_index(seed_num, test_proportion, not_na_index)
}
\arguments{
\item{seed_num}{Seed for reproducibility}

\item{test_proportion}{Proportion of non-NA indices to sample}

\item{not_na_index}{Indices of non-NA values}
}
\value{
Sampled indices to tune k
}
\description{
Sample indices to tune k for kNN imputation
}
