% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{IQR_outlier}
\alias{IQR_outlier}
\title{IQR-based outlier treatment with capping at 5\% / 95\% quantiles}
\usage{
IQR_outlier(df, variables)
}
\arguments{
\item{df}{Data frame with Outliers}

\item{variables}{Features in Data Frame that need outlier Treatment}
}
\value{
Return data frame with variables' outliers treated
}
\description{
IQR-based outlier treatment with capping at 5\% / 95\% quantiles
}
