% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_categorical.R
\name{k_stratified_cv}
\alias{k_stratified_cv}
\title{Split index into k fold}
\usage{
k_stratified_cv(k, df, target_col, train_index, positive, negative)
}
\arguments{
\item{k}{How many fold you want to split}

\item{df}{Data Frame you wanna test with}

\item{target_col}{The column that need to be predict}

\item{train_index}{Training index that need to split in k fold}

\item{positive}{target_col's positive value}

\item{negative}{target_col's negative value}
}
\value{
k stratified index list
}
\description{
Split index into k fold
}
