% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{kNN_Imputation}
\alias{kNN_Imputation}
\title{Perform kNN imputation for a single feature}
\usage{
kNN_Imputation(df, smallest_k, target_feature, df_distance)
}
\arguments{
\item{df}{Data frame that need NA kNN imputation treatment}

\item{smallest_k}{The best k calculate before}

\item{target_feature}{Specific column that need NA kNN imputation}

\item{df_distance}{The euclidean distance matrix made before}
}
\value{
Feature value that successfully imputated
}
\description{
Perform kNN imputation for a single feature
}
