% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{median_imputation}
\alias{median_imputation}
\title{Median imputation for selected columns}
\usage{
median_imputation(df, target_col)
}
\arguments{
\item{df}{The data frame that has NA need median imputation treatment}

\item{target_col}{Columns name that need its NA be treated with median imputation}
}
\value{
Data frame with designated column successfully impute
}
\description{
Median imputation for selected columns
}
