% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_numerical.R
\name{reverse_num}
\alias{reverse_num}
\title{Reverse numerical value into original value, support log, log1p, sqrt, square, default none}
\usage{
reverse_num(x, treatment)
}
\arguments{
\item{x}{The variable that's been or not been transformed}

\item{treatment}{Could be "log","log1p","sqrt","square","none"}
}
\value{
Transform back original value
}
\description{
Reverse numerical value into original value, support log, log1p, sqrt, square, default none
}
