% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_numerical_with_mlr.R
\name{stack_test_mlr}
\alias{stack_test_mlr}
\title{Use Stack Model Make Prediction & Check Its Accuracy with RMSE}
\usage{
stack_test_mlr(
  stack_model,
  model_list,
  df,
  validation_index,
  target_col,
  target_treatment = "none"
)
}
\arguments{
\item{stack_model}{The stack model you build earlier with ensemble models}

\item{model_list}{The list of ensemble model you made ealier}

\item{df}{Data frame that you use to make prediction}

\item{validation_index}{Index used for validation purpose}

\item{target_col}{The feature you want to make prediction with}
}
\value{
Stack model's predict value & actual value & RMSE
}
\description{
Use Stack Model Make Prediction & Check Its Accuracy with RMSE
}
