% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning.R
\name{three_set_partition}
\alias{three_set_partition}
\title{Split data into train, test, and validation with stratified sampling}
\usage{
three_set_partition(
  df,
  target_col,
  train_portion,
  test_portion,
  positive,
  negative
)
}
\arguments{
\item{df}{Data frame that used for partition}

\item{target_col}{Specific column that you don't want to fall in class imbalance (usually column used for prediction)}

\item{train_portion}{Portion of data select for training}

\item{test_portion}{Portion of data select for testing (rest for validation)}

\item{positive}{Target column's positive value}

\item{negative}{Target column's negative value}
}
\value{
List with train_index, test_index, validation_index
}
\description{
Split data into train, test, and validation with stratified sampling
}
