% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{z_score_outlier}
\alias{z_score_outlier}
\title{Z-score based outlier treatment (±3 SD)}
\usage{
z_score_outlier(df, variable)
}
\arguments{
\item{df}{Data frame}

\item{variable}{Character vector of numeric variable names}
}
\value{
Data frame with extreme values capped at mean ± 3*sd
}
\description{
Z-score based outlier treatment (±3 SD)
}
